First draft 1.0.0: separate methods into modules.
