+++ .eslintrc
@@ -8,6 +8,7 @@
     "no-shadow": [0],
     "quotes": [2, "single"],
     "key-spacing": [0],
+    "eqeqeq": [1],
     "curly": [0],
     "camelcase": [0]
   }
