+++ src/epsilon.js
@@ -1,0 +1,7 @@
+'use strict';
+
+// We use `ε`, epsilon, as a stopping criterion when we want to iterate
+// until we're "close enough".
+var epsilon = 0.0001;
+
+module.exports = epsilon;
