+++ src/mode.js
@@ -1,0 +1,58 @@
+'use strict';
+
+// # [mode](http://bit.ly/W5K4Yt)
+//
+// The mode is the number that appears in a list the highest number of times.
+// There can be multiple modes in a list: in the event of a tie, this
+// algorithm will return the most recently seen mode.
+//
+// This implementation is inspired by [science.js](https://github.com/jasondavies/science.js/blob/master/src/stats/mode.js)
+//
+// This runs on `O(n)`, linear time in respect to the array
+function mode(x) {
+
+    // Handle edge cases:
+    // The median of an empty list is null
+    if (x.length === 0) return null;
+    else if (x.length === 1) return x[0];
+
+    // Sorting the array lets us iterate through it below and be sure
+    // that every time we see a new number it's new and we'll never
+    // see the same number twice
+    var sorted = x.slice().sort(function (a, b) { return a - b; });
+
+    // This assumes it is dealing with an array of size > 1, since size
+    // 0 and 1 are handled immediately. Hence it starts at index 1 in the
+    // array.
+    var last = sorted[0],
+        // store the mode as we find new modes
+        value,
+        // store how many times we've seen the mode
+        max_seen = 0,
+        // how many times the current candidate for the mode
+        // has been seen
+        seen_this = 1;
+
+    // end at sorted.length + 1 to fix the case in which the mode is
+    // the highest number that occurs in the sequence. the last iteration
+    // compares sorted[i], which is undefined, to the highest number
+    // in the series
+    for (var i = 1; i < sorted.length + 1; i++) {
+        // we're seeing a new number pass by
+        if (sorted[i] !== last) {
+            // the last number is the new mode since we saw it more
+            // often than the old one
+            if (seen_this > max_seen) {
+                max_seen = seen_this;
+                value = last;
+            }
+            seen_this = 1;
+            last = sorted[i];
+        // if this isn't a new number, it's one more occurrence of
+        // the potential mode
+        } else { seen_this++; }
+    }
+    return value;
+}
+
+module.exports = mode;
