+++ src/z_score.js
@@ -1,0 +1,21 @@
+'use strict';
+
+// # [Z-Score, or Standard Score](http://en.wikipedia.org/wiki/Standard_score)
+//
+// The standard score is the number of standard deviations an observation
+// or datum is above or below the mean. Thus, a positive standard score
+// represents a datum above the mean, while a negative standard score
+// represents a datum below the mean. It is a dimensionless quantity
+// obtained by subtracting the population mean from an individual raw
+// score and then dividing the difference by the population standard
+// deviation.
+//
+// The z-score is only defined if one knows the population parameters;
+// if one only has a sample set, then the analogous computation with
+// sample mean and sample standard deviation yields the
+// Student's t-statistic.
+function z_score(x, mean, standard_deviation) {
+    return (x - mean) / standard_deviation;
+}
+
+module.exports = z_score;
