+++ test/mixin.test.js
@@ -3,7 +3,7 @@ var ss = require('../');
 
 test('mixin', function(t) {
     test('can mix into a single array', function(t) {
-        var even = ss.mixin([2, 4, 6, 8]);
+        var even = ss.mixin(ss, [2, 4, 6, 8]);
         t.equal(even.sum(), 20);
         t.equal(even.mean(), 5);
         t.equal(even.max(), 8);
@@ -13,7 +13,7 @@ test('mixin', function(t) {
     });
 
     test('can mix into Array.prototype', function(t) {
-        ss.mixin();
+        ss.mixin(ss);
         var even = [2, 4, 6, 8];
         t.equal(even.sum(), 20);
         t.equal(even.mean(), 5);
@@ -24,7 +24,7 @@ test('mixin', function(t) {
     });
 
     test('mixins can take arguments', function(t) {
-        ss.mixin();
+        ss.mixin(ss);
         var even = [2, 4, 6, 8];
         t.equal(even.quantile(0.2), 2);
         t.equal(even.quantile(0.8), 8);
